+++ logic/src/main/java/application/Repository/RoleRepository.java
@@ -3,6 +3,6 @@ package application.Repository;
 import application.models.Role;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RoleRepository extends CrudRepository<Role, Integer> {
+public interface RoleRepository extends CrudRepository<Role, Long> {
     Role findById(long roleId);
 }
