+++ logic/src/main/java/application/Repository/ServiceRepository.java
@@ -3,6 +3,6 @@ package application.Repository;
 import application.models.Service;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ServiceRepository extends CrudRepository<Service, Integer> {
+public interface ServiceRepository extends CrudRepository<Service, Long> {
 
 }
